Tie cleanup thread lifetime to containing Device
When a Device is created, it starts a background cleanup thread to
manage releasing interfaces after a delay.  This is logically part of
Device, but was not constrained to the same lifetime.  To make it more
complicated, Device implements Clone, so it can't just stop the thread
in Drop.  If all the Device clones referring to the same underlying
rusb::DeviceHandle were dropped before the USB device was unplugged,
this could cause a panic when the cleanup thread eventually tried to
release interfaces for a closed rusb::DeviceHandle.

To avoid this situation, make InterfaceManager return a new guard object
when it creates the cleanup thread.  Once this guard object is dropped,
the thread will be stopped cleanly and all of the interfaces will be
released.  Device tracks the guard object in a reference-counted field
so it can be dropped when the last clone is dropped.

Fixes #4.