Disable HTTP Keep-Alive
Bridge already filters the Keep-Alive and related headers in requests.
Explicitly disable it in hyper so the framework doesn't keep listening
for a request that will never arrive.

To go with this, also enforce a Connection: close header in responses
from the printer.  This ensures that Bridge isn't passing conflicting
indicators to the client.