Prefer lowest-numbered interfaces
Currently, incoming requests are distributed across all available
IPP-USB interfaces in order. The IPP-USB spec does not prescribe any
particular order for using interfaces, so there is nothing conceptually
wrong with this logic. However, some printers seem to be sensitive to
the order in which interfaces are used.

By always using the lowest-numbered available interface, these printers
seem happy without causing regressions in any of the other test devices
in our lab.  This is achieved by switching out the VecDeque for a
BinaryHeap with a custom Ord that prioritizes ClaimedInterface by lowest
number.  Because BinaryHeap does not provide a non-consuming mutable
iterator, the heap has to be rebuilt each time it is walked to claim or
release interfaces.  This makes operations O(n*lg n) and O(lg n) instead
of O(n) and O(1), but this shouldn't matter because a typical printer
only has a couple of IPP-USB interfaces anyway.